+++ src/main/kotlin/com/coder/gateway/settings/CoderSettings.kt
@@ -65,7 +65,12 @@ open class CoderSettingsState(
     open var enableBinaryDirectoryFallback: Boolean = false,
 
     /**
-     * Controls whether we fall back release.coder.com
+     * Controls whether we verify the cli signature
+     */
+    open var disableSignatureVerification: Boolean = false,
+
+    /**
+     * Controls whether we fall back release.coder.com if signature validation is enabled
      */
     open var fallbackOnCoderForSignatures: Boolean = false,
 
@@ -109,7 +114,7 @@ open class CoderSettingsState(
     // Default version of IDE to display in IDE selection dropdown
     open var defaultIde: String = "",
     // Whether to check for IDE updates.
-    open var checkIDEUpdates: Boolean = true,
+    open var checkIDEUpdates: Boolean = true
 )
 
 /**
@@ -137,7 +142,7 @@ open class CoderSettings(
     // Overrides the default environment (for tests).
     private val env: Environment = Environment(),
     // Overrides the default binary name (for tests).
-    private val binaryName: String? = null,
+    private val binaryName: String? = null
 ) {
     val tls = CoderTLSSettings(state)
 
@@ -159,6 +164,12 @@ open class CoderSettings(
      */
     val enableBinaryDirectoryFallback: Boolean
         get() = state.enableBinaryDirectoryFallback
+
+    /**
+     * Controls whether we verify the cli signature
+     */
+    val disableSignatureVerification: Boolean
+        get() = state.disableSignatureVerification
 
     /**
      * Controls whether we fall back release.coder.com